Linter checks if code blocks are balanced (){}[]
O(n) complexity. O(n) space

Scans each char in the string and adds the corresponding closer symbol
to the stack. When a closer is found, pops the stack to check if the
symbols match. If no, then return false immediatly without scanning the
rest of the string.
If there is a closer but nothing in the stack,then it must be unbalanced

If there are items remaining in the stack after it has been scanned then it can not be balanced