Queue using two stacks passes all tests
If HeadQueue is empty when enqueue is called, push Value there,
otherwise push values to tailQueue.

HeadQueue is the one that is popped() with Dequeue() is called. Should
HeadQueue be empty, pop out every item in the tail queue, reversing it,
and push value into headqueue. This will then be ordered correctly